KAS-wgpu: move parts of flat_round to new common module
